Create Parliament
An important resolution is being discussed in the Parliament. There are N members present in the Parliament out of which X members voted in favour of the resolution and the remaining voted against it.
According to the constitution, a resolution is passed if and only if half or more than half of the members present in the Parliament vote in favour of the resolution.
Determine if the resolution is passed or not.